[wip] Never import generated code, use single entry points for shared modules
