+++ packages/apps/api/tsconfig.json
@@ -1,17 +1,8 @@
 {
+  "extends": "../../../tsconfig.base.json",
   "compilerOptions": {
-    "composite": true,
-    "declaration": true,
-    "declarationMap": true,
-    "esModuleInterop": true,
-    "module": "CommonJS",
-    "moduleResolution": "node",
-    "noImplicitAny": true,
     "outDir": "lib",
-    "rootDir": "src",
-    "sourceMap": true,
-    "strict": true,
-    "target": "es2017"
+    "rootDir": "src"
   },
   "include": ["src", "test"],
   "references": [
@@ -20,6 +11,9 @@
     },
     {
       "path": "../../shared/auth0-client"
+    },
+    {
+      "path": "../../shared/identity-common"
     }
   ]
 }
