+++ packages/apps/auth0-actions/tsconfig.json
@@ -4,5 +4,5 @@
     "outDir": "lib",
     "rootDir": "src"
   },
-  "include": ["src", "test"]
+  "include": ["src", "tests"]
 }
